requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0